+++ CHANGELOG.md
@@ -54,6 +54,8 @@ Renamed "Additional Actions" sidebar to "Quick Access" and updated the icon used
 
 Resolve an edge case when running notebooks where a false-failure would be detected that halts notebook execution.
 
+Fixed an issue when launching IDL on Mac where ENVI crashes because "LANG" is missing from the environment
+
 ## 3.2.3 September 2023
 
 Re-work cancellation to handle cases where the worker threads were too busy to get the messages that work needs to be stopped. If you now use a very large file (take "slicer3.pro" from the IDL distribution) events are interrupted as expected.